Added dropper sub
Session wrtier called without callback uses dropper as callback

Optimised the static file serving with reordering of eof checking. Save
an extra callback now.

Also fixed a performance issue with attempting to open a file with no
read permissions. A stat is now done beforehand, and access rights
checked before the open is attempted.

the stat on existing open files is not done on the fh for a slight
improvement in performance.