Rework of writer stack.. again
With no callback porvided, the session dropper is used as the callback.

When about to be called, the cb is checked for equality with dropper.
If they are the same, no argument is used for dropper. This would
potentially force the session to close.

Otherwise the argument provided (or __SUB__) is used as cb argument.

If there is actually an error, the dropper is called with a true
argument, forcing the session to drop and clean up