Added better multi HTTP method routing
Seperate methods can now be added to a route in individual
defines_route calls.

This means that additional method testing might not be required in the
rout handler, as the method type is part of the match.

For example, a login form, using GET to deliver the HTML page/form to
client, and the POST to process the response, avoids extra testing of
the method:

	define_route GET=>  "/login"=>sub {...};
	define_route POST=> "/login"=>sub {...}

instead of
	define_route "GET|POST"=>"/login"=>sub {
		my $rex=$_[1];
		given($rex->[method_]){
			when("GET"){
				...
			}
			when("POST"){
				...
			}
		}
	}

Each uri endpoint with negative matching methdods (ie any method) is
added to the end of the dispatch table.

If the URI matches but is of the incorrect method, a unsported method
error is returned.

If the URI doesn't match, it is then handled by the default handler,
which is a not found error