Added/fixed redirection support
Rex now has redirection subs as sugar to unclude the correct status code
and set the location header.

The an internal redirect sub will rewrite the uri in the request and
then redispatch it through the dispatcher

rex_site_url is also added to create a url that is within the current
site/group.