Reworked state cookie processing
The state_uuid and and state_json are are now seperate middlewares.

The state_cookie only handles decoding/encoding base64 data in to a
cookie and state field.

To prevent a cookie from being sent, make the state field undef before
sending
To expire/delete a cookie, set state field to empty string

Otherwise the values in the state field will reencoded on output

TODO: Still need to set an expiring on the actual cookie...