Refined Header rendering.
Clawing back some performance

Using perls string/number typing, headers constants are now indexes into
a name array. When the user specifes a string header, this evaluates to
a 0, which results in an undefined value from the name array. Using the
defined or operator, the string used as the index, which is actaully a
name is the result.