Resolved memory and lost rex bug
Added weaken reference to session in rex.

IF a client requests are resource/range and doesn't read its socket
entirely then the dropper would drop the session. This would leave the
callback active to read mor data however.

A check has been added to only service the rex if it is indeed still
active