Temp fix of compression middleware
For small read sizes the extra file reads required where not being
executed.  The callbacks where not being called when the
scratch/compression output was not filled.

This needs another look as the code will over flow slow links as the
callback is not tied to when the data is sent