Fixed internal redirect bug
Internal redirect was not passing the correct number of arguments on the
newly found route. This caused a strange error in the static file
middleware which I was using at the time.

Made some notes in the Middleware.pm file reguarding calling conventions
and important values for the middleware processing chain.