Improved middleware, routing and serializing
Middleware is now expected to be an array ref containing  inner, outer
and name values.

inner and outer must be code refs, but are not required.

if a route has a code ref used directly as middleware, it is wrapped to
make it call next. The user sub only has access to message variables,
but can call rex_error etc to alter the direction of message flow.

Sites/groups not have a controller field. This is either a object
reference or a package name. If a route is specified with a
string/scalar  as middleware, it is used as a method name against the
controller package/object. The returned value is to be a  middleware
array ref or code ref suitable for further processing.