More reporting to static middleware
When meta cache was disabled, file descriptors where never closed. This
has been fixed by adding a cached_ flag on every entry.

If it is set, the entry was stored in the cache on opening, and is left
for the cache to close it.

Still need to remove cache items on errors and also on actual file
changes.