Adjusted middleware code status
A default code of -1 is now used at the start of processing. middleware
can inspect and conditionally execute from this code. Eg chained static
middle ware can test if the code for -1 or 404 to determin if it should
process.

If the code is less then 0 at serialization time, it is forces to 200.

Updated static middleware to make use of this update.

TODO: Need to keep an eye on internal redirects updating the code value.
Currenly it is not reset to -1 on an internal redirect.