Reworked static indexs and preencodings
Preencodings need to be specified in a encoding=>extension hash.
The accept encoding in request is checked and the extension is looked
up. This is appended to the original file path to find a new pre encoded
file.

File cache no stores headers in user field.

indexs files are checked when enabled. The first file found that matches
is used.