Migrated to errrorware
Middleware now  has an optional error sub which is called when
underlying IO stream is in an error state.

The error sub gives the inner/outer ware pairs a change to release any
resources.

The errorware is stored in the route entry and is execute by a parser.
The parser is called with 0 arguments when an error occured in the
file handle.  This is translated into a call to the errorware head
instead of innerware/outerware head