Static file changes
An undefined html_root spec is now treated as file relative instead of
process relative

Changes to setting of status in static middleware. This needs testing.
The ideas is if the code is not defined, then set any errors from static
file reading. If a status code is already set then no not override.

Static not has option of addition prefix stripping to allow chaining of
static html root locations