Normalized prefixing and route adding
Prefixes of site are simply text. Normally you would end the prefix with
a slash to end the 'path' component, so a route starts a new path
component.

However, if it does't end in a slash, the the currnet path component is
modified.

The root of the site hierarchy on the server always has a prefix of '/'.

For a route to match the prefix, simply use an empty string as the path

Removed hash route addition. Use a loop over $site->add_route instead