Added template support to static middleware
Passing in a template option, which is a regex matcher, a match will
result in the path of the file being placed in payload, instead of
contents of the file. The status is also set to undef

This allows the next middleware  (a template middleware) to process the
path stored in payload, render it to generate response payload.